Fix migration FK constraint: account_id must be BIGINT to match accounts.id
accounts.id is BIGINT; INT account_id columns cause errno 150 on FK creation.

+++ flaskapp/migrations/add_location_based_budgeting.sql
@@ -165,7 +165,7 @@ CREATE TABLE IF NOT EXISTS campaign_location_targets (
 
     -- Campaign
     campaign_id INT NOT NULL,
-    account_id INT NOT NULL,
+    account_id BIGINT NOT NULL,
 
     -- Location targeting
     location_criterion_id VARCHAR(50) NOT NULL COMMENT 'Google Ads location criterion ID',
